pyproject.toml
[tool.poetry]
name = "django-bridger"
version = "0.10.0"
description = "The bridge between a Django Backend and a Javascript based Frontend"
authors = ["Christopher Wittlinger <c.wittlinger@intellineers.com>"]

packages = [
    { include = "bridger" },
]

[tool.poetry.dependencies]
python = "3.8.*"
django = "3.0.*"
django-filter = "2.2.*"
djangorestframework = "3.11.*"
djangorestframework-simplejwt = "4.3.*"
django-fsm = "2.7.*"
requests = "2.24.*"
channels = "2.4.*"
channels_redis = {version = "3.0.*", extras=["cryptography"]}
celery = {version = "4.4.*", extras=["redis"]}
django-celery-beat = "2.0.*"
psycopg2-binary = "2.8.*"
pillow = "7.2.*"
numpy = "1.19.*"
pandas = "1.0.*"
jsonschema = "3.2.*"
plotly = "4.8.*"
django-simple-history = "2.11.*"
markdown = "3.2.*"
markdown-blockdiag = "0.7.*"
toml = "0.10.*"
python-decouple = "3.3"
python-slugify = "*"

selenium = "3.141.*"
pytest-selenium = "1.17.*"
termcolor = "^1.1.0"

[tool.poetry.dev-dependencies]
# Testing dependencies
pytest = "5.2.*"
coverage = "5.2.*"
pytest-django = "3.9.*"
pytest-cov = "2.10.*"
pytest-mock = "3.1.*"
pytest-factoryboy = "2.0.*"
pytest-asyncio = "0.10.*"
pytest-flake8 = "1.0.*"
pytest-isort = "1.0.*"
tox = "3.16.*"
devtools = {version="0.5.*", extras=["pygments"]}
pynvim = "0.4.*"

# Documentation dependencies
mkdocs = "1.1.*"
mkdocs-material = "5.4.*"
markdown-include = "0.5.*"
pymdown-extensions = "7.1.*"

# Formatting and Linting dependencies
black = "19.10b0"
flake8 = "3.8.*"
isort = "5.0.4"

# Test Implementation dependencies
boto3 = "1.14.*"
factory-boy = "2.12.*"
django-extensions = "3.0.*"
django-cors-headers = "3.4.*"
django-storages = "1.9.*"

# Misc. dependencies
colorlog = "4.1.*"
dj-database-url = "0.5.*"
jupyter = "1.0.*"

# Security Vulnerabilities
safety = "1.9.*"

[tool.coverage.run]
omit = [
    "tests/*",
    "setup.py",
    ".tox/*",
    "*/migrations/*",
    "manage.py",

    "*/metadata_validator.py",
]

[build-system]
requires = ["poetry_core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

